RNA-seq data from Blackmore et al. 2017

Dataset from The effect of upper-respiratory infection on transcriptomic changes in the CNS by Blackmore et al. (2017) (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5544260/) used as data for the Bioconductor intro and RNA-seq lessons and WSBIM1207 courses. The two variables are of class SummarizedExperiment.

Usage

data("GSE96870_intro")
data("GSE96870_intro_ranges")

Details

For details on how the data was prepared, see https://github.com/Bioconductor/bioconductor-teaching/tree/master/data/GSE96870.

Examples

data("GSE96870_intro")
GSE96870_intro
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, table, tapply,
#>     union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
#> class: RangedSummarizedExperiment 
#> dim: 1474 22 
#> metadata(0):
#> assays(1): counts
#> rownames(1474): Asl Apod ... Pbx1 Rgs4
#> rowData names(11): gene ENTREZID ... phenotype_description
#>   hsapiens_homolog_associated_gene_name
#> colnames(22): GSM2545336 GSM2545337 ... GSM2545363 GSM2545380
#> colData names(10): title geo_accession ... tissue mouse
rowData(GSE96870_intro)
#> DataFrame with 1474 rows and 11 columns
#>                gene    ENTREZID                product       gbkey
#>         <character> <character>            <character> <character>
#> Asl             Asl      109900 argininosuccinate ly..        mRNA
#> Apod           Apod       11815 apolipoprotein D, tr..        mRNA
#> Cyp2d22     Cyp2d22       56448 cytochrome P450, fam..        mRNA
#> Klk6           Klk6       19144 kallikrein related-p..        mRNA
#> Fcrls         Fcrls       80891 Fc receptor-like S, ..        mRNA
#> ...             ...         ...                    ...         ...
#> Lrrc52       Lrrc52      240899 leucine rich repeat ..        mRNA
#> Rxrg           Rxrg       20183 retinoid X receptor ..        mRNA
#> Lmx1a         Lmx1a      110648 LIM homeobox transcr..        mRNA
#> Pbx1           Pbx1       18514 pre B cell leukemia ..        mRNA
#> Rgs4           Rgs4       19736 regulator of G-prote..        mRNA
#>         external_gene_name    ensembl_gene_id external_synonym chromosome_name
#>                <character>        <character>      <character>     <character>
#> Asl                    Asl ENSMUSG00000025533    2510006M18Rik               5
#> Apod                  Apod ENSMUSG00000022548               NA              16
#> Cyp2d22            Cyp2d22 ENSMUSG00000061740             2D22              15
#> Klk6                  Klk6 ENSMUSG00000050063             Bssp               7
#> Fcrls                Fcrls ENSMUSG00000015852    2810439C17Rik               3
#> ...                    ...                ...              ...             ...
#> Lrrc52              Lrrc52 ENSMUSG00000040485    4930413P14Rik               1
#> Rxrg                  Rxrg ENSMUSG00000015843            Nr2b3               1
#> Lmx1a                Lmx1a ENSMUSG00000026686           Lmx1.1               1
#> Pbx1                  Pbx1 ENSMUSG00000052534    2310056B04Rik               1
#> Rgs4                  Rgs4 ENSMUSG00000038530           ESTM48               1
#>           gene_biotype  phenotype_description
#>            <character>            <character>
#> Asl     protein_coding abnormal circulating..
#> Apod    protein_coding abnormal lipid homeo..
#> Cyp2d22 protein_coding abnormal skin morpho..
#> Klk6    protein_coding abnormal cytokine le..
#> Fcrls   protein_coding decreased CD8-positi..
#> ...                ...                    ...
#> Lrrc52  protein_coding abnormal sperm physi..
#> Rxrg    protein_coding abnormal bone minera..
#> Lmx1a   protein_coding abnormal bony labyri..
#> Pbx1    protein_coding abnormal adrenal gla..
#> Rgs4    protein_coding abnormal touch/ noci..
#>         hsapiens_homolog_associated_gene_name
#>                                   <character>
#> Asl                                       ASL
#> Apod                                     APOD
#> Cyp2d22                                CYP2D6
#> Klk6                                     KLK6
#> Fcrls                                   FCRL4
#> ...                                       ...
#> Lrrc52                                 LRRC52
#> Rxrg                                     RXRG
#> Lmx1a                                   LMX1A
#> Pbx1                                     PBX1
#> Rgs4                                     RGS4

data("GSE96870_intro_ranges")
GSE96870_intro_ranges
#> class: RangedSummarizedExperiment 
#> dim: 1474 22 
#> metadata(0):
#> assays(1): counts
#> rownames(1474): Asl Apod ... Pbx1 Rgs4
#> rowData names(11): gene ENTREZID ... phenotype_description
#>   hsapiens_homolog_associated_gene_name
#> colnames(22): GSM2545336 GSM2545337 ... GSM2545363 GSM2545380
#> colData names(10): title geo_accession ... tissue mouse
rowRanges(GSE96870_intro_ranges)
#> GRanges object with 1474 ranges and 11 metadata columns:
#>           seqnames              ranges strand |        gene    ENTREZID
#>              <Rle>           <IRanges>  <Rle> | <character> <character>
#>       Asl        5 130024208-130024330      - |         Asl      109900
#>      Apod       16   31314552-31314808      - |        Apod       11815
#>   Cyp2d22       15   82380078-82380260      - |     Cyp2d22       56448
#>      Klk6        7   43824544-43824595      + |        Klk6       19144
#>     Fcrls        3   87263678-87263765      - |       Fcrls       80891
#>       ...      ...                 ...    ... .         ...         ...
#>    Lrrc52        1 167466093-167466780      - |      Lrrc52      240899
#>      Rxrg        1 167598362-167598800      + |        Rxrg       20183
#>     Lmx1a        1 167688300-167688991      + |       Lmx1a      110648
#>      Pbx1        1 168431314-168432169      - |        Pbx1       18514
#>      Rgs4        1 169747480-169747642      - |        Rgs4       19736
#>                          product       gbkey external_gene_name
#>                      <character> <character>        <character>
#>       Asl argininosuccinate ly..        mRNA                Asl
#>      Apod apolipoprotein D, tr..        mRNA               Apod
#>   Cyp2d22 cytochrome P450, fam..        mRNA            Cyp2d22
#>      Klk6 kallikrein related-p..        mRNA               Klk6
#>     Fcrls Fc receptor-like S, ..        mRNA              Fcrls
#>       ...                    ...         ...                ...
#>    Lrrc52 leucine rich repeat ..        mRNA             Lrrc52
#>      Rxrg retinoid X receptor ..        mRNA               Rxrg
#>     Lmx1a LIM homeobox transcr..        mRNA              Lmx1a
#>      Pbx1 pre B cell leukemia ..        mRNA               Pbx1
#>      Rgs4 regulator of G-prote..        mRNA               Rgs4
#>              ensembl_gene_id external_synonym chromosome_name   gene_biotype
#>                  <character>      <character>     <character>    <character>
#>       Asl ENSMUSG00000025533    2510006M18Rik               5 protein_coding
#>      Apod ENSMUSG00000022548             <NA>              16 protein_coding
#>   Cyp2d22 ENSMUSG00000061740             2D22              15 protein_coding
#>      Klk6 ENSMUSG00000050063             Bssp               7 protein_coding
#>     Fcrls ENSMUSG00000015852    2810439C17Rik               3 protein_coding
#>       ...                ...              ...             ...            ...
#>    Lrrc52 ENSMUSG00000040485    4930413P14Rik               1 protein_coding
#>      Rxrg ENSMUSG00000015843            Nr2b3               1 protein_coding
#>     Lmx1a ENSMUSG00000026686           Lmx1.1               1 protein_coding
#>      Pbx1 ENSMUSG00000052534    2310056B04Rik               1 protein_coding
#>      Rgs4 ENSMUSG00000038530           ESTM48               1 protein_coding
#>            phenotype_description hsapiens_homolog_associated_gene_name
#>                      <character>                           <character>
#>       Asl abnormal circulating..                                   ASL
#>      Apod abnormal lipid homeo..                                  APOD
#>   Cyp2d22 abnormal skin morpho..                                CYP2D6
#>      Klk6 abnormal cytokine le..                                  KLK6
#>     Fcrls decreased CD8-positi..                                 FCRL4
#>       ...                    ...                                   ...
#>    Lrrc52 abnormal sperm physi..                                LRRC52
#>      Rxrg abnormal bone minera..                                  RXRG
#>     Lmx1a abnormal bony labyri..                                 LMX1A
#>      Pbx1 abnormal adrenal gla..                                  PBX1
#>      Rgs4 abnormal touch/ noci..                                  RGS4
#>   -------
#>   seqinfo: 182 sequences from an unspecified genome; no seqlengths