Coronavirus (COVID-19) Cases

These data were downloaded from Novel Coronavirus COVID-19 (2019-nCoV) Data Repository by Johns Hopkins CSSE (https://github.com/CSSEGISandData/COVID-19):

This is the data repository for the 2019 Novel Coronavirus Visual Dashboard operated by the Johns Hopkins University Center for Systems Science and Engineering (JHU CSSE). Also, Supported by ESRI Living Atlas Team and the Johns Hopkins University Applied Physics Lab (JHU APL).

The `covid19_cases.csv()`, `covid19_deaths.csv()` and `covid19_recoevered.csv()` functions return the path to a comma-separated file containing confirmed cases, deaths and recovered cases over time for a certain number of coutries/regions. See example below for details.

Usage

covid19_cases.csv()
covid19_deaths.csv()
covid19_recovered.csv()

Examples

cv <- readr::read_csv(covid19_cases.csv())
#> Rows: 279 Columns: 580
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (2): Province_State, Country_Region
#> dbl (578): Lat, Long, 1/22/20, 1/23/20, 1/24/20, 1/25/20, 1/26/20, 1/27/20, ...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
## dates (fomat: month/day/year)
names(cv)[-(1:4)]
#>   [1] "1/22/20"  "1/23/20"  "1/24/20"  "1/25/20"  "1/26/20"  "1/27/20" 
#>   [7] "1/28/20"  "1/29/20"  "1/30/20"  "1/31/20"  "2/1/20"   "2/2/20"  
#>  [13] "2/3/20"   "2/4/20"   "2/5/20"   "2/6/20"   "2/7/20"   "2/8/20"  
#>  [19] "2/9/20"   "2/10/20"  "2/11/20"  "2/12/20"  "2/13/20"  "2/14/20" 
#>  [25] "2/15/20"  "2/16/20"  "2/17/20"  "2/18/20"  "2/19/20"  "2/20/20" 
#>  [31] "2/21/20"  "2/22/20"  "2/23/20"  "2/24/20"  "2/25/20"  "2/26/20" 
#>  [37] "2/27/20"  "2/28/20"  "2/29/20"  "3/1/20"   "3/2/20"   "3/3/20"  
#>  [43] "3/4/20"   "3/5/20"   "3/6/20"   "3/7/20"   "3/8/20"   "3/9/20"  
#>  [49] "3/10/20"  "3/11/20"  "3/12/20"  "3/13/20"  "3/14/20"  "3/15/20" 
#>  [55] "3/16/20"  "3/17/20"  "3/18/20"  "3/19/20"  "3/20/20"  "3/21/20" 
#>  [61] "3/22/20"  "3/23/20"  "3/24/20"  "3/25/20"  "3/26/20"  "3/27/20" 
#>  [67] "3/28/20"  "3/29/20"  "3/30/20"  "3/31/20"  "4/1/20"   "4/2/20"  
#>  [73] "4/3/20"   "4/4/20"   "4/5/20"   "4/6/20"   "4/7/20"   "4/8/20"  
#>  [79] "4/9/20"   "4/10/20"  "4/11/20"  "4/12/20"  "4/13/20"  "4/14/20" 
#>  [85] "4/15/20"  "4/16/20"  "4/17/20"  "4/18/20"  "4/19/20"  "4/20/20" 
#>  [91] "4/21/20"  "4/22/20"  "4/23/20"  "4/24/20"  "4/25/20"  "4/26/20" 
#>  [97] "4/27/20"  "4/28/20"  "4/29/20"  "4/30/20"  "5/1/20"   "5/2/20"  
#> [103] "5/3/20"   "5/4/20"   "5/5/20"   "5/6/20"   "5/7/20"   "5/8/20"  
#> [109] "5/9/20"   "5/10/20"  "5/11/20"  "5/12/20"  "5/13/20"  "5/14/20" 
#> [115] "5/15/20"  "5/16/20"  "5/17/20"  "5/18/20"  "5/19/20"  "5/20/20" 
#> [121] "5/21/20"  "5/22/20"  "5/23/20"  "5/24/20"  "5/25/20"  "5/26/20" 
#> [127] "5/27/20"  "5/28/20"  "5/29/20"  "5/30/20"  "5/31/20"  "6/1/20"  
#> [133] "6/2/20"   "6/3/20"   "6/4/20"   "6/5/20"   "6/6/20"   "6/7/20"  
#> [139] "6/8/20"   "6/9/20"   "6/10/20"  "6/11/20"  "6/12/20"  "6/13/20" 
#> [145] "6/14/20"  "6/15/20"  "6/16/20"  "6/17/20"  "6/18/20"  "6/19/20" 
#> [151] "6/20/20"  "6/21/20"  "6/22/20"  "6/23/20"  "6/24/20"  "6/25/20" 
#> [157] "6/26/20"  "6/27/20"  "6/28/20"  "6/29/20"  "6/30/20"  "7/1/20"  
#> [163] "7/2/20"   "7/3/20"   "7/4/20"   "7/5/20"   "7/6/20"   "7/7/20"  
#> [169] "7/8/20"   "7/9/20"   "7/10/20"  "7/11/20"  "7/12/20"  "7/13/20" 
#> [175] "7/14/20"  "7/15/20"  "7/16/20"  "7/17/20"  "7/18/20"  "7/19/20" 
#> [181] "7/20/20"  "7/21/20"  "7/22/20"  "7/23/20"  "7/24/20"  "7/25/20" 
#> [187] "7/26/20"  "7/27/20"  "7/28/20"  "7/29/20"  "7/30/20"  "7/31/20" 
#> [193] "8/1/20"   "8/2/20"   "8/3/20"   "8/4/20"   "8/5/20"   "8/6/20"  
#> [199] "8/7/20"   "8/8/20"   "8/9/20"   "8/10/20"  "8/11/20"  "8/12/20" 
#> [205] "8/13/20"  "8/14/20"  "8/15/20"  "8/16/20"  "8/17/20"  "8/18/20" 
#> [211] "8/19/20"  "8/20/20"  "8/21/20"  "8/22/20"  "8/23/20"  "8/24/20" 
#> [217] "8/25/20"  "8/26/20"  "8/27/20"  "8/28/20"  "8/29/20"  "8/30/20" 
#> [223] "8/31/20"  "9/1/20"   "9/2/20"   "9/3/20"   "9/4/20"   "9/5/20"  
#> [229] "9/6/20"   "9/7/20"   "9/8/20"   "9/9/20"   "9/10/20"  "9/11/20" 
#> [235] "9/12/20"  "9/13/20"  "9/14/20"  "9/15/20"  "9/16/20"  "9/17/20" 
#> [241] "9/18/20"  "9/19/20"  "9/20/20"  "9/21/20"  "9/22/20"  "9/23/20" 
#> [247] "9/24/20"  "9/25/20"  "9/26/20"  "9/27/20"  "9/28/20"  "9/29/20" 
#> [253] "9/30/20"  "10/1/20"  "10/2/20"  "10/3/20"  "10/4/20"  "10/5/20" 
#> [259] "10/6/20"  "10/7/20"  "10/8/20"  "10/9/20"  "10/10/20" "10/11/20"
#> [265] "10/12/20" "10/13/20" "10/14/20" "10/15/20" "10/16/20" "10/17/20"
#> [271] "10/18/20" "10/19/20" "10/20/20" "10/21/20" "10/22/20" "10/23/20"
#> [277] "10/24/20" "10/25/20" "10/26/20" "10/27/20" "10/28/20" "10/29/20"
#> [283] "10/30/20" "10/31/20" "11/1/20"  "11/2/20"  "11/3/20"  "11/4/20" 
#> [289] "11/5/20"  "11/6/20"  "11/7/20"  "11/8/20"  "11/9/20"  "11/10/20"
#> [295] "11/11/20" "11/12/20" "11/13/20" "11/14/20" "11/15/20" "11/16/20"
#> [301] "11/17/20" "11/18/20" "11/19/20" "11/20/20" "11/21/20" "11/22/20"
#> [307] "11/23/20" "11/24/20" "11/25/20" "11/26/20" "11/27/20" "11/28/20"
#> [313] "11/29/20" "11/30/20" "12/1/20"  "12/2/20"  "12/3/20"  "12/4/20" 
#> [319] "12/5/20"  "12/6/20"  "12/7/20"  "12/8/20"  "12/9/20"  "12/10/20"
#> [325] "12/11/20" "12/12/20" "12/13/20" "12/14/20" "12/15/20" "12/16/20"
#> [331] "12/17/20" "12/18/20" "12/19/20" "12/20/20" "12/21/20" "12/22/20"
#> [337] "12/23/20" "12/24/20" "12/25/20" "12/26/20" "12/27/20" "12/28/20"
#> [343] "12/29/20" "12/30/20" "12/31/20" "1/1/21"   "1/2/21"   "1/3/21"  
#> [349] "1/4/21"   "1/5/21"   "1/6/21"   "1/7/21"   "1/8/21"   "1/9/21"  
#> [355] "1/10/21"  "1/11/21"  "1/12/21"  "1/13/21"  "1/14/21"  "1/15/21" 
#> [361] "1/16/21"  "1/17/21"  "1/18/21"  "1/19/21"  "1/20/21"  "1/21/21" 
#> [367] "1/22/21"  "1/23/21"  "1/24/21"  "1/25/21"  "1/26/21"  "1/27/21" 
#> [373] "1/28/21"  "1/29/21"  "1/30/21"  "1/31/21"  "2/1/21"   "2/2/21"  
#> [379] "2/3/21"   "2/4/21"   "2/5/21"   "2/6/21"   "2/7/21"   "2/8/21"  
#> [385] "2/9/21"   "2/10/21"  "2/11/21"  "2/12/21"  "2/13/21"  "2/14/21" 
#> [391] "2/15/21"  "2/16/21"  "2/17/21"  "2/18/21"  "2/19/21"  "2/20/21" 
#> [397] "2/21/21"  "2/22/21"  "2/23/21"  "2/24/21"  "2/25/21"  "2/26/21" 
#> [403] "2/27/21"  "2/28/21"  "3/1/21"   "3/2/21"   "3/3/21"   "3/4/21"  
#> [409] "3/5/21"   "3/6/21"   "3/7/21"   "3/8/21"   "3/9/21"   "3/10/21" 
#> [415] "3/11/21"  "3/12/21"  "3/13/21"  "3/14/21"  "3/15/21"  "3/16/21" 
#> [421] "3/17/21"  "3/18/21"  "3/19/21"  "3/20/21"  "3/21/21"  "3/22/21" 
#> [427] "3/23/21"  "3/24/21"  "3/25/21"  "3/26/21"  "3/27/21"  "3/28/21" 
#> [433] "3/29/21"  "3/30/21"  "3/31/21"  "4/1/21"   "4/2/21"   "4/3/21"  
#> [439] "4/4/21"   "4/5/21"   "4/6/21"   "4/7/21"   "4/8/21"   "4/9/21"  
#> [445] "4/10/21"  "4/11/21"  "4/12/21"  "4/13/21"  "4/14/21"  "4/15/21" 
#> [451] "4/16/21"  "4/17/21"  "4/18/21"  "4/19/21"  "4/20/21"  "4/21/21" 
#> [457] "4/22/21"  "4/23/21"  "4/24/21"  "4/25/21"  "4/26/21"  "4/27/21" 
#> [463] "4/28/21"  "4/29/21"  "4/30/21"  "5/1/21"   "5/2/21"   "5/3/21"  
#> [469] "5/4/21"   "5/5/21"   "5/6/21"   "5/7/21"   "5/8/21"   "5/9/21"  
#> [475] "5/10/21"  "5/11/21"  "5/12/21"  "5/13/21"  "5/14/21"  "5/15/21" 
#> [481] "5/16/21"  "5/17/21"  "5/18/21"  "5/19/21"  "5/20/21"  "5/21/21" 
#> [487] "5/22/21"  "5/23/21"  "5/24/21"  "5/25/21"  "5/26/21"  "5/27/21" 
#> [493] "5/28/21"  "5/29/21"  "5/30/21"  "5/31/21"  "6/1/21"   "6/2/21"  
#> [499] "6/3/21"   "6/4/21"   "6/5/21"   "6/6/21"   "6/7/21"   "6/8/21"  
#> [505] "6/9/21"   "6/10/21"  "6/11/21"  "6/12/21"  "6/13/21"  "6/14/21" 
#> [511] "6/15/21"  "6/16/21"  "6/17/21"  "6/18/21"  "6/19/21"  "6/20/21" 
#> [517] "6/21/21"  "6/22/21"  "6/23/21"  "6/24/21"  "6/25/21"  "6/26/21" 
#> [523] "6/27/21"  "6/28/21"  "6/29/21"  "6/30/21"  "7/1/21"   "7/2/21"  
#> [529] "7/3/21"   "7/4/21"   "7/5/21"   "7/6/21"   "7/7/21"   "7/8/21"  
#> [535] "7/9/21"   "7/10/21"  "7/11/21"  "7/12/21"  "7/13/21"  "7/14/21" 
#> [541] "7/15/21"  "7/16/21"  "7/17/21"  "7/18/21"  "7/19/21"  "7/20/21" 
#> [547] "7/21/21"  "7/22/21"  "7/23/21"  "7/24/21"  "7/25/21"  "7/26/21" 
#> [553] "7/27/21"  "7/28/21"  "7/29/21"  "7/30/21"  "7/31/21"  "8/1/21"  
#> [559] "8/2/21"   "8/3/21"   "8/4/21"   "8/5/21"   "8/6/21"   "8/7/21"  
#> [565] "8/8/21"   "8/9/21"   "8/10/21"  "8/11/21"  "8/12/21"  "8/13/21" 
#> [571] "8/14/21"  "8/15/21"  "8/16/21"  "8/17/21"  "8/18/21"  "8/19/21" 
## Countries/Regions
unique(cv[[2]])
#>   [1] "Afghanistan"                      "Albania"                         
#>   [3] "Algeria"                          "Andorra"                         
#>   [5] "Angola"                           "Antigua and Barbuda"             
#>   [7] "Argentina"                        "Armenia"                         
#>   [9] "Australia"                        "Austria"                         
#>  [11] "Azerbaijan"                       "Bahamas"                         
#>  [13] "Bahrain"                          "Bangladesh"                      
#>  [15] "Barbados"                         "Belarus"                         
#>  [17] "Belgium"                          "Belize"                          
#>  [19] "Benin"                            "Bhutan"                          
#>  [21] "Bolivia"                          "Bosnia and Herzegovina"          
#>  [23] "Botswana"                         "Brazil"                          
#>  [25] "Brunei"                           "Bulgaria"                        
#>  [27] "Burkina Faso"                     "Burma"                           
#>  [29] "Burundi"                          "Cabo Verde"                      
#>  [31] "Cambodia"                         "Cameroon"                        
#>  [33] "Canada"                           "Central African Republic"        
#>  [35] "Chad"                             "Chile"                           
#>  [37] "China"                            "Colombia"                        
#>  [39] "Comoros"                          "Congo (Brazzaville)"             
#>  [41] "Congo (Kinshasa)"                 "Costa Rica"                      
#>  [43] "Cote d'Ivoire"                    "Croatia"                         
#>  [45] "Cuba"                             "Cyprus"                          
#>  [47] "Czechia"                          "Denmark"                         
#>  [49] "Diamond Princess"                 "Djibouti"                        
#>  [51] "Dominica"                         "Dominican Republic"              
#>  [53] "Ecuador"                          "Egypt"                           
#>  [55] "El Salvador"                      "Equatorial Guinea"               
#>  [57] "Eritrea"                          "Estonia"                         
#>  [59] "Eswatini"                         "Ethiopia"                        
#>  [61] "Fiji"                             "Finland"                         
#>  [63] "France"                           "Gabon"                           
#>  [65] "Gambia"                           "Georgia"                         
#>  [67] "Germany"                          "Ghana"                           
#>  [69] "Greece"                           "Grenada"                         
#>  [71] "Guatemala"                        "Guinea"                          
#>  [73] "Guinea-Bissau"                    "Guyana"                          
#>  [75] "Haiti"                            "Holy See"                        
#>  [77] "Honduras"                         "Hungary"                         
#>  [79] "Iceland"                          "India"                           
#>  [81] "Indonesia"                        "Iran"                            
#>  [83] "Iraq"                             "Ireland"                         
#>  [85] "Israel"                           "Italy"                           
#>  [87] "Jamaica"                          "Japan"                           
#>  [89] "Jordan"                           "Kazakhstan"                      
#>  [91] "Kenya"                            "Kiribati"                        
#>  [93] "Korea, South"                     "Kosovo"                          
#>  [95] "Kuwait"                           "Kyrgyzstan"                      
#>  [97] "Laos"                             "Latvia"                          
#>  [99] "Lebanon"                          "Lesotho"                         
#> [101] "Liberia"                          "Libya"                           
#> [103] "Liechtenstein"                    "Lithuania"                       
#> [105] "Luxembourg"                       "MS Zaandam"                      
#> [107] "Madagascar"                       "Malawi"                          
#> [109] "Malaysia"                         "Maldives"                        
#> [111] "Mali"                             "Malta"                           
#> [113] "Marshall Islands"                 "Mauritania"                      
#> [115] "Mauritius"                        "Mexico"                          
#> [117] "Micronesia"                       "Moldova"                         
#> [119] "Monaco"                           "Mongolia"                        
#> [121] "Montenegro"                       "Morocco"                         
#> [123] "Mozambique"                       "Namibia"                         
#> [125] "Nepal"                            "Netherlands"                     
#> [127] "New Zealand"                      "Nicaragua"                       
#> [129] "Niger"                            "Nigeria"                         
#> [131] "North Macedonia"                  "Norway"                          
#> [133] "Oman"                             "Pakistan"                        
#> [135] "Palau"                            "Panama"                          
#> [137] "Papua New Guinea"                 "Paraguay"                        
#> [139] "Peru"                             "Philippines"                     
#> [141] "Poland"                           "Portugal"                        
#> [143] "Qatar"                            "Romania"                         
#> [145] "Russia"                           "Rwanda"                          
#> [147] "Saint Kitts and Nevis"            "Saint Lucia"                     
#> [149] "Saint Vincent and the Grenadines" "Samoa"                           
#> [151] "San Marino"                       "Sao Tome and Principe"           
#> [153] "Saudi Arabia"                     "Senegal"                         
#> [155] "Serbia"                           "Seychelles"                      
#> [157] "Sierra Leone"                     "Singapore"                       
#> [159] "Slovakia"                         "Slovenia"                        
#> [161] "Solomon Islands"                  "Somalia"                         
#> [163] "South Africa"                     "South Sudan"                     
#> [165] "Spain"                            "Sri Lanka"                       
#> [167] "Sudan"                            "Summer Olympics 2020"            
#> [169] "Suriname"                         "Sweden"                          
#> [171] "Switzerland"                      "Syria"                           
#> [173] "Taiwan*"                          "Tajikistan"                      
#> [175] "Tanzania"                         "Thailand"                        
#> [177] "Timor-Leste"                      "Togo"                            
#> [179] "Trinidad and Tobago"              "Tunisia"                         
#> [181] "Turkey"                           "US"                              
#> [183] "Uganda"                           "Ukraine"                         
#> [185] "United Arab Emirates"             "United Kingdom"                  
#> [187] "Uruguay"                          "Uzbekistan"                      
#> [189] "Vanuatu"                          "Venezuela"                       
#> [191] "Vietnam"                          "West Bank and Gaza"              
#> [193] "Yemen"                            "Zambia"                          
#> [195] "Zimbabwe"                        
## Provice/States
unique(cv[[1]])
#>  [1] NA                                            
#>  [2] "Australian Capital Territory"                
#>  [3] "New South Wales"                             
#>  [4] "Northern Territory"                          
#>  [5] "Queensland"                                  
#>  [6] "South Australia"                             
#>  [7] "Tasmania"                                    
#>  [8] "Victoria"                                    
#>  [9] "Western Australia"                           
#> [10] "Alberta"                                     
#> [11] "British Columbia"                            
#> [12] "Diamond Princess"                            
#> [13] "Grand Princess"                              
#> [14] "Manitoba"                                    
#> [15] "New Brunswick"                               
#> [16] "Newfoundland and Labrador"                   
#> [17] "Northwest Territories"                       
#> [18] "Nova Scotia"                                 
#> [19] "Nunavut"                                     
#> [20] "Ontario"                                     
#> [21] "Prince Edward Island"                        
#> [22] "Quebec"                                      
#> [23] "Repatriated Travellers"                      
#> [24] "Saskatchewan"                                
#> [25] "Yukon"                                       
#> [26] "Anhui"                                       
#> [27] "Beijing"                                     
#> [28] "Chongqing"                                   
#> [29] "Fujian"                                      
#> [30] "Gansu"                                       
#> [31] "Guangdong"                                   
#> [32] "Guangxi"                                     
#> [33] "Guizhou"                                     
#> [34] "Hainan"                                      
#> [35] "Hebei"                                       
#> [36] "Heilongjiang"                                
#> [37] "Henan"                                       
#> [38] "Hong Kong"                                   
#> [39] "Hubei"                                       
#> [40] "Hunan"                                       
#> [41] "Inner Mongolia"                              
#> [42] "Jiangsu"                                     
#> [43] "Jiangxi"                                     
#> [44] "Jilin"                                       
#> [45] "Liaoning"                                    
#> [46] "Macau"                                       
#> [47] "Ningxia"                                     
#> [48] "Qinghai"                                     
#> [49] "Shaanxi"                                     
#> [50] "Shandong"                                    
#> [51] "Shanghai"                                    
#> [52] "Shanxi"                                      
#> [53] "Sichuan"                                     
#> [54] "Tianjin"                                     
#> [55] "Tibet"                                       
#> [56] "Unknown"                                     
#> [57] "Xinjiang"                                    
#> [58] "Yunnan"                                      
#> [59] "Zhejiang"                                    
#> [60] "Faroe Islands"                               
#> [61] "Greenland"                                   
#> [62] "French Guiana"                               
#> [63] "French Polynesia"                            
#> [64] "Guadeloupe"                                  
#> [65] "Martinique"                                  
#> [66] "Mayotte"                                     
#> [67] "New Caledonia"                               
#> [68] "Reunion"                                     
#> [69] "Saint Barthelemy"                            
#> [70] "Saint Pierre and Miquelon"                   
#> [71] "St Martin"                                   
#> [72] "Wallis and Futuna"                           
#> [73] "Aruba"                                       
#> [74] "Bonaire, Sint Eustatius and Saba"            
#> [75] "Curacao"                                     
#> [76] "Sint Maarten"                                
#> [77] "Cook Islands"                                
#> [78] "Anguilla"                                    
#> [79] "Bermuda"                                     
#> [80] "British Virgin Islands"                      
#> [81] "Cayman Islands"                              
#> [82] "Channel Islands"                             
#> [83] "Falkland Islands (Malvinas)"                 
#> [84] "Gibraltar"                                   
#> [85] "Isle of Man"                                 
#> [86] "Montserrat"                                  
#> [87] "Saint Helena, Ascension and Tristan da Cunha"
#> [88] "Turks and Caicos Islands"