TidyTuesday's English Education Data

These two files describe the 'Educational attainment of young people in English towns' from the the UK Office for National Statistics. It was explored in the July 2023 article "Why do children and young people in smaller towns do better academically than those in larger towns?".

Two files are available:

english_education.csv was prepared as part of the Tidy Tuesday series on Educational attainment of young people in English towns (2024-01-23). The page also describes the variables.

edu_income_eprivation_and_educational_attainment.csv was downloaded from the UK Office for National Statistics and converted from xls to csv. The variables in this table are also described on the page linked above.

Usage

english_education_files()

Value

character(2) with file names.

References

"Why do children and young people in smaller towns do better academically than those in larger towns?".

Tidy Tuesday data from 2024-01-23: https://github.com/rfordatascience/tidytuesday/blob/master/data/2024/2024-01-23/readme.md

Author

Laurent Gatto

Examples

english_education_files()
#> [1] "/home/runner/work/_temp/Library/rWSBIM1207/extdata/edu_income_eprivation_and_educational_attainment.csv"        
#> [2] "https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2024/2024-01-23/english_education.csv"