Road accidents data

Number of killed, seriously injured, slightly injured and uninjured victims of road accidents, by age group, type of user, sex and various characteristics of the accident in Belgium in 20222.

Nombre de tués, blessés graves, blessés légers et victimes indemnes d'accidents de la route, par classe d'âges, genre d'usager, sexe et diverses caractéristiques de l'accident en Belgique en 2022.

These are publicly available data downloaded on 16 June 2023 from the Belgian gouvernment open data page. For more details on these data, see https://statbel.fgov.be/fr/themes/mobilite/circulation/accidents-de-la-circulation

Usage

road_accidents_be_2022.rds()
road_accidents_be_meta.csv()

Examples

readRDS(road_accidents_be_2022.rds())
#> # A tibble: 82,876 × 43
#>    DT_DAY     DT_HOUR CD_DAY_OF_WEEK TX_DAY_OF_WEEK_DESCR_FR
#>    <date>       <dbl>          <dbl> <chr>                  
#>  1 2022-12-13       7              2 Mardi                  
#>  2 2022-11-13       9              7 Dimanche               
#>  3 2022-11-21      17              1 Lundi                  
#>  4 2022-10-16      14              7 Dimanche               
#>  5 2022-12-30      16              5 Vendredi               
#>  6 2022-11-29      13              2 Mardi                  
#>  7 2022-11-02      17              3 Mercredi               
#>  8 2022-10-16      14              7 Dimanche               
#>  9 2022-10-21      19              5 Vendredi               
#> 10 2022-12-17      15              6 Samedi                 
#> # ℹ 82,866 more rows
#> # ℹ 39 more variables: TX_DAY_OF_WEEK_DESCR_NL <chr>, MS_VICT <dbl>,
#> #   MS_VIC_OK <dbl>, MS_SLY_INJ <dbl>, MS_SERLY_INJ <dbl>,
#> #   MS_DEAD_30_DAYS <dbl>, CD_BUILD_UP_AREA <dbl>,
#> #   TX_BUILD_UP_AREA_DESCR_NL <chr>, TX_BUILD_UP_AREA_DESCR_FR <chr>,
#> #   CD_VICT_TYPE <dbl>, TX_VICT_TYPE_DESCR_FR <chr>,
#> #   TX_VICT_TYPE_DESCR_NL <chr>, CD_ROAD_USER_TYPE <chr>, …
read.csv(road_accidents_be_meta.csv())
#>                         NAME                                           LABEL
#> 1                     DT_DAY                                             Day
#> 2                    DT_HOUR                                     Time of day
#> 3             CD_DAY_OF_WEEK                                     Day of week
#> 4    TX_DAY_OF_WEEK_DESCR_FR                                Day of the month
#> 5    TX_DAY_OF_WEEK_DESCR_NL                                Day of the month
#> 6                     MS_VCT                               Number of victims
#> 7                  MS_VIC_OK                             Number of uninjured
#> 8                 MS_SLY_INJ                      Number of slightly injured
#> 9               MS_SERLY_INJ                      Number of severely injured
#> 10               MS_MORY_INJ                      Number of mortally injured
#> 11                   MS_DEAD                                  Number of dead
#> 12           MS_DEAD_30_DAYS                          Number of dead 30 days
#> 13               CD_VCT_TYPE                                    Business key
#> 14      TX_VCT_TYPE_DESCR_NL Victime type (driver, passenger, other victime)
#> 15      TX_VCT_TYPE_DESCR_FR Victime type (driver, passenger, other victime)
#> 16          CD_ROAD_USR_TYPE                                    Business key
#> 17 TX_ROAD_USR_TYPE_DESCR_NL                                  Road user type
#> 18 TX_ROAD_USR_TYPE_DESCR_FR                                  Road user type
#> 19              CD_ROAD_TYPE                                    Business key
#> 20     TX_ROAD_TYPE_DESCR_NL                                       Road type
#> 21     TX_ROAD_TYPE_DESCR_FR                                       Road type
#> 22             CD_LIGHT_COND                                    Business key
#> 23    TX_LIGHT_COND_DESCR_NL                                Light conditions
#> 24    TX_LIGHT_COND_DESCR_FR                                Light conditions
#> 25              CD_COLL_TYPE                                    Business key
#> 26     TX_COLL_TYPE_DESCR_NL                                  Collision type
#> 27     TX_COLL_TYPE_DESCR_FR                                  Collision type
#> 28          CD_BUILD_UP_AREA                                    Business key
#> 29 TX_BUILD_UP_AREA_DESCR_NL                                   Build up area
#> 30 TX_BUILD_UP_AREA_DESCR_FR                                   Build up area
#> 31                CD_AGE_CLS                                    Business key
#> 32       TX_AGE_CLS_DESCR_NL                                       Age class
#> 33       TX_AGE_CLS_DESCR_FR                                       Age class
#> 34           CD_MUNTY_REFNIS                 REFNIS code of the municipality
#> 35         TX_MUNTY_DESCR_NL                      Municipality name in Dutch
#> 36         TX_MUNTY_DESCR_FR                     Municipality name in French
#> 37            CD_DSTR_REFNIS                     REFNIS code of the district
#> 38      TX_ADM_DSTR_DESCR_NL           Administrative district name in Dutch
#> 39      TX_ADM_DSTR_DESCR_FR          Administrative district name in French
#> 40            CD_PROV_REFNIS                     REFNIS code of the province
#> 41          TX_PROV_DESCR_NL                          Province name in Dutch
#> 42          TX_PROV_DESCR_FR                         Province name in French
#> 43             CD_RGN_REFNIS                       REFNIS code of the region
#> 44           TX_RGN_DESCR_NL                            Region name in Dutch
#> 45           TX_RGN_DESCR_FR                           Region name in French
#> 46                    CD_SEX                                    Business key
#> 47           TX_SEX_DESCR_NL                                             Sex
#> 48           TX_SEX_DESCR_FR                                             Sex