KEM RNA-Seq data

This data is composed of three files, namely kem_counts.tsv, kem_counts2.tsv, containing counts data and the annotation file kem_annot.tsv containing the annotations the KEM samples. Both files are encoded as tab-separtated sheets and can be found with the kem.tsv() function.

Credit: the data have been generated by Mr Kevin Missault.

Format

The RNA-Seq count data for 13 ENSEMBL transcripts (ref) and 16 KEM samples is encoded as:

ref

ENSEMBL transcript identifiers.

KEM182-01

Expression counts for all genes in sample KEM182-01.

KEM182-02

Expression counts for all genes in sample KEM182-02.

KEM182-03

Expression counts for all genes in sample KEM182-03.

KEM182-04

Expression counts for all genes in sample KEM182-04.

KEM182-05

Expression counts for all genes in sample KEM182-05.

KEM182-06

Expression counts for all genes in sample KEM182-06.

KEM182-07

Expression counts for all genes in sample KEM182-07.

KEM182-08

Expression counts for all genes in sample KEM182-08.

KEM182-09

Expression counts for all genes in sample KEM182-09.

KEM182-10

Expression counts for all genes in sample KEM182-10.

KEM182-11

Expression counts for all genes in sample KEM182-11.

KEM182-12

Expression counts for all genes in sample KEM182-12.

KEM182-13

Expression counts for all genes in sample KEM182-13.

KEM182-14

Expression counts for all genes in sample KEM182-14.

KEM182-15

Expression counts for all genes in sample KEM182-15.

KEM182-16

Expression counts for all genes in sample KEM182-16.

The kem_counts2.tsv file contains counts data for 4774 features.

The annotation contains the following variables for the 16 observations:

sample_id

Sample identifier.

jurkat

A character (yes or no) defining of the cells a Jurkat cells.

cell_type

Jurkat cell type (A or B).

treatment

Treatment: either none or stimulated.

Examples

kem.tsv()
#> [1] "/home/runner/work/_temp/Library/rWSBIM1207/extdata/kem_counts.tsv"
#> [2] "/home/runner/work/_temp/Library/rWSBIM1207/extdata/kem_annot.tsv" 
kem2.tsv()
#> [1] "/home/runner/work/_temp/Library/rWSBIM1207/extdata/kem_counts2.tsv"
#> [2] "/home/runner/work/_temp/Library/rWSBIM1207/extdata/kem_annot.tsv"  
kem3.tsv()
#> [1] "/home/runner/work/_temp/Library/rWSBIM1207/extdata/kem_counts3a.tsv"
#> [2] "/home/runner/work/_temp/Library/rWSBIM1207/extdata/kem_counts3b.tsv"
#> [3] "/home/runner/work/_temp/Library/rWSBIM1207/extdata/kem_annot.tsv"